MicroserviceLibrary/pv_infer/requirements.txt
python-consul==1.1.0
matplotlib==3.9.0
numpy==1.26.4
redis==5.0.4
torch
torchvision
torchaudio